+++ checkpoint/orbax/checkpoint/experimental/v1/_src/loading/loading.py
@@ -20,16 +20,12 @@ from typing import Any
 
 from absl import logging
 from etils import epath
-from orbax.checkpoint._src.checkpointers import async_checkpointer
-from orbax.checkpoint._src.handlers import composite_checkpoint_handler
-from orbax.checkpoint._src.handlers import handler_registration as legacy_handler_registration
 from orbax.checkpoint._src.serialization import type_handlers
 from orbax.checkpoint.experimental.v1._src.context import context as context_lib
-from orbax.checkpoint.experimental.v1._src.handlers import compatibility as handler_compatibility
-from orbax.checkpoint.experimental.v1._src.handlers import composite_handler
 import orbax.checkpoint.experimental.v1._src.handlers.global_registration  # pylint: disable=unused-import
 from orbax.checkpoint.experimental.v1._src.layout import checkpoint_layout
 from orbax.checkpoint.experimental.v1._src.layout import registry as layout_registry
+from orbax.checkpoint.experimental.v1._src.loading import validation
 from orbax.checkpoint.experimental.v1._src.metadata import types as metadata_types
 from orbax.checkpoint.experimental.v1._src.path import format_utils
 from orbax.checkpoint.experimental.v1._src.path import types as path_types
@@ -49,18 +45,6 @@ def _standardize_abstract_checkpointables(abstract_checkpointables):
   if isinstance(abstract_checkpointables, CheckpointMetadata):
     return abstract_checkpointables.metadata
   return abstract_checkpointables
-
-
-def _validate_abstract_checkpointables(abstract_checkpointables):
-  if abstract_checkpointables is None:
-    return
-  if (
-      provided_reserved_keys := abstract_checkpointables.keys()
-      & format_utils.RESERVED_CHECKPOINTABLE_KEYS
-  ):
-    raise ValueError(
-        f'Provided reserved checkpointable keys: {provided_reserved_keys}.'
-    )
 
 
 def load_pytree(
@@ -238,7 +222,7 @@ def _load_checkpointables_impl(
   abstract_checkpointables = _standardize_abstract_checkpointables(
       abstract_checkpointables
   )
-  _validate_abstract_checkpointables(abstract_checkpointables)
+  validation.validate_abstract_checkpointables(abstract_checkpointables)
 
   async def _load() -> dict[str, Any]:
     load_awaitable = await layout.load(abstract_checkpointables)
@@ -284,61 +268,3 @@ def load_checkpointables_async(
   """Loads a checkpointables asynchronously. Not yet implemented."""
   del path, abstract_checkpointables
   raise NotImplementedError('Asynchronous loading is not yet supported.')
-
-
-def get_v0_checkpointer_and_args(
-    path: path_types.Path,
-    abstract_checkpointables: dict[str, Any] | None,
-    *,
-    context: context_lib.Context,
-) -> tuple[
-    async_checkpointer.AsyncCheckpointer,
-    composite_checkpoint_handler.CompositeArgs,
-]:
-  """Construct V0 Checkpointer and Args for loading."""
-  _validate_abstract_checkpointables(abstract_checkpointables)
-  abstract_checkpointables = abstract_checkpointables or {}
-
-  # pylint: disable=protected-access
-  handlers = composite_handler.CompositeHandler(
-      context.checkpointables_options.registry
-  ).get_handlers_for_load(path, abstract_checkpointables)
-  # pylint: enable=protected-access
-  if not abstract_checkpointables:
-    abstract_checkpointables = {
-        name: None
-        for name in handlers.keys()
-        if name not in format_utils.RESERVED_CHECKPOINTABLE_KEYS
-        and (path / name).exists()
-    }
-
-  compatibility_handlers = {
-      name: handler_compatibility.get_compatibility_handler(handler)
-      for name, handler in handlers.items()
-  }
-  legacy_handler_registry = (
-      legacy_handler_registration.DefaultCheckpointHandlerRegistry()
-  )
-  for name, handler in compatibility_handlers.items():
-    legacy_handler_registry.add(name, handler_compatibility.Args, handler)
-  composite_options = composite_checkpoint_handler.CompositeOptions(
-      async_options=context.async_options.v0(),
-      file_options=context.file_options.v0(),
-      multiprocessing_options=context.multiprocessing_options.v0(),
-      temporary_path_class=context.file_options.temporary_path_class,
-  )
-  ckptr = async_checkpointer.AsyncCheckpointer(
-      composite_checkpoint_handler.CompositeCheckpointHandler(
-          handler_registry=legacy_handler_registry,
-          composite_options=composite_options,
-      ),
-      async_options=context.async_options.v0(),
-      multiprocessing_options=context.multiprocessing_options.v0(),
-      file_options=context.file_options.v0(),
-      temporary_path_class=context.file_options.temporary_path_class,
-  )
-  args = composite_checkpoint_handler.CompositeArgs(**{
-      name: handler_compatibility.Args(checkpointable)
-      for name, checkpointable in abstract_checkpointables.items()
-  })
-  return ckptr, args