+++ checkpoint/orbax/checkpoint/experimental/v1/_src/training/checkpointer.py
@@ -23,7 +23,7 @@ from etils import epy
 from orbax.checkpoint import checkpoint_manager
 from orbax.checkpoint.experimental.v1._src.context import context as context_lib
 import orbax.checkpoint.experimental.v1._src.handlers.global_registration  # pylint: disable=unused-import
-from orbax.checkpoint.experimental.v1._src.loading import loading
+from orbax.checkpoint.experimental.v1._src.loading import v0_compatibility as v0_loading_utils
 from orbax.checkpoint.experimental.v1._src.metadata import loading as metadata_loading
 from orbax.checkpoint.experimental.v1._src.metadata import types as metadata_types
 from orbax.checkpoint.experimental.v1._src.path import format_utils
@@ -427,7 +427,7 @@ class Checkpointer(epy.ContextManager):
   ) -> dict[str, Any]:
     """Loads a set of checkpointables at the given step."""
     step = self._resolve_existing_checkpoint(step).step
-    checkpointer, args = loading.get_v0_checkpointer_and_args(
+    checkpointer, args = v0_loading_utils.get_v0_checkpointer_and_args(
         self.directory / self._step_name_format.build_name(step),
         abstract_checkpointables,
         context=context_lib.get_context(),